chore: Rename as index on content of folder
