+++ client/src/Components/Button/index.spec.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Button from '~/Components/Button/Button';
+import Button from '~/Components/Button';
 
 describe('<Button />', () => {
   /**
