+++ client/src/Components/Copyright/index.spec.tsx
@@ -3,11 +3,11 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Divider from '~/Components/Divider/Divider';
+import Copyright from '~/Components/Copyright';
 
 describe('<Button />', () => {
   it('should render same with snapshot', () => {
-    const { container } = render(<Divider />);
+    const { container } = render(<Copyright>저작권 관련 내용</Copyright>);
     expect(container).toMatchSnapshot();
   });
 });
