+++ client/src/Components/Copyright/index.tsx
