+++ client/src/Components/Divider/index.spec.tsx
@@ -3,11 +3,11 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Input from '~/Components/Input/Input';
+import Divider from '~/Components/Divider';
 
 describe('<Button />', () => {
   it('should render same with snapshot', () => {
-    const { container } = render(<Input />);
+    const { container } = render(<Divider />);
     expect(container).toMatchSnapshot();
   });
 });
