+++ client/src/Components/InputHelp/index.spec.tsx
@@ -3,11 +3,11 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Copyright from '~/Components/Copyright/Copyright';
+import InputHelp from '~/Components/InputHelp';
 
 describe('<Button />', () => {
   it('should render same with snapshot', () => {
-    const { container } = render(<Copyright>저작권 관련 내용</Copyright>);
+    const { container } = render(<InputHelp>비밀번호가 너무 짧아요</InputHelp>);
     expect(container).toMatchSnapshot();
   });
 });
